src/main/java/shop/Shop.java
package shop;
import java.util.ArrayList;

public class Shop {

    ArrayList<ISell> stock;

    public Shop() {
        this.stock = new ArrayList<>();
    }

    public ArrayList<ISell> getStock() {
        return stock;
    }

    public void setStock(ArrayList<ISell> stock) {
        this.stock = stock;
    }

    public void addToStock(ISell item){
        this.stock.add(item);
    }

    public void removeFromStock(ISell item){
        this.stock.remove(item);
    }

    public double calculateTotalPotentialProfit(){
        double total  =
        stock
                .stream()
                .mapToDouble(ISell::calculateMarkup)
                .sum();
        return total;
    }
}
